Feat/css fanciness (#5)
* Initial Commit

Reorganizing some things to make it a little cleaner on my end, and detailing the basic concepts I want to go for. Ideally the general look is going to be of an old hi-fi section.

* Basic ideas

General color setup, will work on alignment and better looks after lunch

* Animations

When you add a new row, and when you hit the calculate button, there's a real neat "unfolding component" animation now. Up next is the piano key styling on the radio buttons. (Gonna take some heavy overriding)

* Casette Length Styling

After a great deal of fiddling with the css for radio buttons, I've got nice fancy piano key looking buttons. I even used the linear-gradient function to make them look "pressed in".  Up next is the woodgrain to replace that drab wood color, font choice, and mobile responsiveness. Mobile responsiveness should only require overriding some margins/padding, everything else should fit, regardless of display width.

* Final Background Changes

Added some fonts and woodgrain! Now all that's left is mobile responsiveness, but that's a problem for another branch and another day.
