+++ casette-calculator/public/index.html
@@ -22,6 +22,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <link href="https://fonts.googleapis.com/css?family=Share+Tech+Mono|ZCOOL+QingKe+HuangYou" rel="stylesheet">
     <title>React App</title>
   </head>
   <body>
