+++ casette-calculator/src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import InputForm from '../src/components/InputForm';
+import InputForm from './components/InputForm/InputForm';
 import './App.css';
-import LengthPicker from './components/LengthPicker';
+import LengthPicker from './components/LengthPicker/LengthPicker';
 
 class App extends Component {
   render() {
