--- casette-calculator/src/components/InputForm.js
@@ -1,23 +1,0 @@
-import React, { Component } from 'react';
-import InputRow from '../components/InputRow.js';
-import SubmitButton from '../components/SubmitButton.js';
-
-
-class InputForm extends Component 
-{
-    render()
-    {
-        
-        return(
-            <div className="inputform">                
-                <InputRow>
-                
-                </InputRow>
-                <SubmitButton>
-
-                </SubmitButton>
-            </div>
-        )
-    }
-}
-export default InputForm;
