+++ casette-calculator/src/components/InputForm/InputForm.css
@@ -1,0 +1,51 @@
+.inputform
+{
+    background-image: linear-gradient(#f2f2f2, silver,#f2f2f2);
+    padding: 1em;
+}
+.inputform .inputrow
+{
+    background-color: black;
+    color: #4ac3b7;
+    animation-name: foldIn;
+    animation-iteration-count: 1;
+    animation-duration: 3s;
+    animation-fill-mode: forwards;
+}
+.inputform .inputrow *:not(.newRow)
+{
+    background-color: transparent;
+    animation-name: foldIn;
+    animation-iteration-count: 1;
+    animation-duration: 3s;
+    animation-fill-mode: forwards;
+}
+.newRow
+{
+    animation-name: align-fix;
+    animation-duration: 3s;
+    animation-fill-mode: forwards;
+    animation-iteration-count: 1;
+}
+@keyframes foldIn
+{
+    0%
+    {
+        height:0em;
+    }
+    100%
+    {
+        height:1.7em;
+    } 
+}
+@keyframes align-fix
+{
+    0%
+    {
+        margin-top: .2em;
+    }
+    100%
+    {
+        margin-top: 2.2em;
+    }
+}
