+++ casette-calculator/src/components/InputMinutes/InputMinutes.css
@@ -1,0 +1,8 @@
+.minute
+{
+    background-color: black;
+    color: #4ac3b7;
+    width:3em;
+    border:none;
+    text-align: center;
+}
