+++ casette-calculator/src/components/InputRow/InputRow.css
@@ -1,0 +1,11 @@
+.newRow
+{
+    margin-top: 2.2em;
+    margin-right: .1em;
+    float:right;
+    border-radius: 90%;
+}
+.inputrow
+{
+    padding-left: 48%;
+}
