+++ casette-calculator/src/components/InputSeconds/InputSeconds.css
@@ -1,0 +1,8 @@
+.second
+{
+    background-color: black;
+    color: #4ac3b7;
+    width: 3em;
+    border:none;
+    text-align: center;
+}
