+++ casette-calculator/src/components/InputSeconds/InputSeconds.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import { increment } from '../../scripts/IncrementScripts';
+import "./InputSeconds.css";
+
+class InputSeconds extends Component 
+{
+    render()
+    {
+        return(
+        <input type="number" min="0" max="60" className="second" 
+        onChange={increment}>
+        </input>)
+    }
+}
+export default InputSeconds;
