+++ casette-calculator/src/components/LengthPicker/LengthPicker.css
@@ -1,0 +1,30 @@
+.lengthselector
+{
+    /* the radio buttons inside need to look like "piano key" buttons*/
+    background-color: #f2f2f2;
+    padding-left: 2%;
+}
+input[type=radio]
+{
+    background-color: ivory;
+    width: 24.6%;
+    height: 10em;
+    appearance: none;
+    outline:ivory groove;
+    margin-right: -2%;
+}
+input[type=radio]:checked
+{
+    outline: #f6f6e8 inset;
+    background-image: linear-gradient(ivory 50%,#eaeadc);
+}
+label
+{
+    position:relative;
+    left:-12%;
+}
+label[for="110"]
+{
+    position: relative;
+    left: -11%;
+}
