+++ casette-calculator/src/components/SubmitButton/SubmitButton.css
@@ -1,0 +1,5 @@
+.calculationButton
+{
+    /*neat gradient on the background to look like it's a shiny button*/
+    border-radius: 100%;
+}
